+++ src/client/components/ProjectSidebar.tsx
@@ -1,6 +1,16 @@
 /** @jsx jsx */
 import React, { useState, Fragment } from "react";
-import { Box, Button, Flex, Heading, jsx, Spinner, Styled, ThemeUIStyleObject } from "theme-ui";
+import {
+  Box,
+  Button,
+  Flex,
+  Heading,
+  jsx,
+  Spinner,
+  Styled,
+  Text,
+  ThemeUIStyleObject
+} from "theme-ui";
 
 import {
   UintArrays,
@@ -13,7 +23,7 @@ import {
   IStaticMetadata,
   LockedDistricts
 } from "../../shared/entities";
-import { DistrictGeoJSON, DistrictsGeoJSON } from "../types";
+import { DistrictGeoJSON, DistrictsGeoJSON, SavingState } from "../types";
 import {
   allGeoUnitIndices,
   areAnyGeoUnitsSelected,
@@ -57,10 +67,15 @@ const style: ThemeUIStyleObject = {
     flexDirection: "column",
     flexShrink: 0,
     height: "100%",
-    minWidth: "415px",
+    minWidth: "430px",
     position: "relative",
     color: "gray.8",
     zIndex: 200
+  },
+  table: {
+    width: "calc(100% - 16px)",
+    mx: 2,
+    mb: 2
   },
   tooltip: {
     position: "absolute",
@@ -81,6 +96,7 @@ const style: ThemeUIStyleObject = {
     position: "sticky",
     top: "0",
     zIndex: 2,
+    userSelect: "none",
     "&::after": {
       height: "1px",
       content: "''",
@@ -100,7 +116,7 @@ const style: ThemeUIStyleObject = {
     fontWeight: "body",
     color: "gray.8",
     fontSize: 1,
-    px: 2,
+    p: 2,
     textAlign: "left",
     verticalAlign: "bottom"
   },
@@ -116,6 +132,11 @@ const style: ThemeUIStyleObject = {
   },
   blankValue: {
     color: "gray.2"
+  },
+  lockButton: {
+    p: "6px",
+    display: "inline-block",
+    lineHeight: "0"
   }
 };
 
@@ -128,7 +149,8 @@ const ProjectSidebar = ({
   selectedDistrictId,
   selectedGeounits,
   geoUnitHierarchy,
-  lockedDistricts
+  lockedDistricts,
+  saving
 }: {
   readonly project?: IProject;
   readonly geojson?: DistrictsGeoJSON;
@@ -138,16 +160,17 @@ const ProjectSidebar = ({
   readonly selectedGeounits: GeoUnits;
   readonly geoUnitHierarchy?: GeoUnitHierarchy;
   readonly lockedDistricts: LockedDistricts;
+  readonly saving: SavingState;
 } & LoadingProps) => {
   return (
     <Flex sx={style.sidebar}>
-      <SidebarHeader selectedGeounits={selectedGeounits} isLoading={isLoading} />
+      <SidebarHeader selectedGeounits={selectedGeounits} isLoading={isLoading} saving={saving} />
       <Box sx={{ overflowY: "auto", flex: 1 }}>
-        <Styled.table sx={{ width: "calc(100% - 16px)", mx: 2, mb: 2 }}>
+        <Styled.table sx={style.table}>
           <thead>
             <Styled.tr>
               <Styled.th sx={style.th}>
-                <Tooltip content="Unique number for this district">
+                <Tooltip content="District number">
                   <span>Number</span>
                 </Tooltip>
               </Styled.th>
@@ -167,7 +190,7 @@ const ProjectSidebar = ({
                 </Tooltip>
               </Styled.th>
               <Styled.th sx={{ ...style.th, ...style.number }}>
-                <Tooltip content="Compactness score">
+                <Tooltip content="Compactness score (Polsby-Popper)">
                   <span>Comp.</span>
                 </Tooltip>
               </Styled.th>
@@ -185,6 +208,7 @@ const ProjectSidebar = ({
                 selectedGeounits={selectedGeounits}
                 geoUnitHierarchy={geoUnitHierarchy}
                 lockedDistricts={lockedDistricts}
+                saving={saving}
               />
             )}
           </tbody>
@@ -196,9 +220,11 @@ const ProjectSidebar = ({
 
 const SidebarHeader = ({
   selectedGeounits,
-  isLoading
+  isLoading,
+  saving
 }: {
   readonly selectedGeounits: GeoUnits;
+  readonly saving: SavingState;
 } & LoadingProps) => {
   return (
     <Flex sx={style.header}>
@@ -207,31 +233,56 @@ const SidebarHeader = ({
           Districts
         </Heading>
       </Flex>
-      {isLoading ? (
+      {isLoading || saving === "saving" ? (
         <Flex sx={{ alignItems: "center", justifyContent: "center" }}>
           <Spinner variant="spinner.small" />
         </Flex>
       ) : areAnyGeoUnitsSelected(selectedGeounits) ? (
         <Flex sx={{ variant: "header.right" }}>
-          <Button
-            variant="circularSubtle"
-            sx={{ mr: "2" }}
-            onClick={() => {
-              store.dispatch(clearSelectedGeounits());
-            }}
+          <Tooltip
+            placement="top-start"
+            content={
+              <span>
+                <strong>Cancel changes</strong> to revert to your previously saved map
+              </span>
+            }
           >
-            Cancel
-          </Button>
-          <Button
-            variant="circular"
-            onClick={() => {
-              store.dispatch(updateDistrictsDefinition());
-            }}
+            <Button
+              variant="circularSubtle"
+              sx={{ mr: "2" }}
+              onClick={() => {
+                store.dispatch(clearSelectedGeounits(true));
+              }}
+            >
+              Cancel
+            </Button>
+          </Tooltip>
+          <Tooltip
+            placement="top-start"
+            content={
+              <span>
+                <strong>Accept changes</strong> to save your map
+              </span>
+            }
           >
-            <Icon name="check" />
-            Accept
-          </Button>
+            <Button
+              variant="circular"
+              onClick={() => {
+                store.dispatch(updateDistrictsDefinition());
+              }}
+            >
+              <Icon name="check" />
+              Accept
+            </Button>
+          </Tooltip>
         </Flex>
+      ) : saving === "saved" ? (
+        <Tooltip placement="top-start" content={<span>Your map is saved</span>}>
+          <Flex sx={{ display: "flex", color: "gray.3", alignItems: "center", userSelect: "none" }}>
+            <Icon name="check-circle" size={1.1} />
+            <Text sx={{ fontSize: 1, ml: 1 }}>Saved</Text>
+          </Flex>
+        </Tooltip>
       ) : null}
     </Flex>
   );
@@ -383,19 +434,20 @@ const SidebarRow = ({
               </span>
             }
           >
-            <span onClick={toggleLocked} sx={{ display: "inline-block", lineHeight: "0" }}>
+            <Button variant="icon" onClick={toggleLocked} sx={style.lockButton}>
               <Icon name="lock-locked" color="#131f28" size={0.75} />
-            </span>
+            </Button>
           </Tooltip>
         ) : (
           <Tooltip content="Lock this district">
-            <span
+            <Button
+              variant="icon"
               style={{ visibility: isHovered ? "visible" : "hidden" }}
               onClick={toggleLocked}
-              sx={{ display: "inline-block", lineHeight: "0" }}
+              sx={style.lockButton}
             >
               <Icon name="lock-unlocked" size={0.75} />
-            </span>
+            </Button>
           </Tooltip>
         )}
       </Styled.td>
@@ -470,6 +522,7 @@ interface SidebarRowsProps {
   readonly selectedGeounits: GeoUnits;
   readonly geoUnitHierarchy: GeoUnitHierarchy;
   readonly lockedDistricts: LockedDistricts;
+  readonly saving: SavingState;
 }
 
 const SidebarRows = ({