+++ src/client/reducers/projectData.ts
@@ -4,8 +4,8 @@ import { getType } from "typesafe-actions";
 import { Action } from "../actions";
 import {
   updateDistrictsDefinition,
-  updateDistrictsDefinitionFailure,
   updateDistrictsDefinitionSuccess,
+  updateDistrictsDefinitionFailure,
   projectFetch,
   projectFetchSuccess,
   projectFetchFailure,
@@ -18,11 +18,16 @@ import {
 import { clearSelectedGeounits } from "../actions/districtDrawing";
 import { ProjectState, initialProjectState } from "./project";
 import { resetProjectState } from "../actions/root";
-import { DynamicProjectData, StaticProjectData } from "../types";
+import { DistrictsGeoJSON, DynamicProjectData, StaticProjectData } from "../types";
 import { Resource } from "../resource";
 
-import { allGeoUnitIndices, assignGeounitsToDistrict } from "../functions";
-import { fetchProjectData, patchProject } from "../api";
+import {
+  allGeoUnitIndices,
+  assignGeounitsToDistrict,
+  showActionFailedToast,
+  showResourceFailedToast
+} from "../functions";
+import { fetchProjectData, fetchProjectGeoJson, patchProject } from "../api";
 import { fetchAllStaticData } from "../s3";
 
 export type ProjectDataState = {
@@ -72,15 +77,18 @@ const projectDataReducer: LoopReducer<ProjectState, Action> = (
             resource: action.payload
           }
         },
-        Cmd.action(clearSelectedGeounits())
+        Cmd.action(clearSelectedGeounits(false))
       );
     case getType(projectFetchFailure):
-      return {
-        ...state,
-        projectData: {
-          errorMessage: action.payload
-        }
-      };
+      return loop(
+        {
+          ...state,
+          projectData: {
+            errorMessage: action.payload
+          }
+        },
+        Cmd.run(showActionFailedToast)
+      );
     case getType(projectDataFetch):
       return loop(
         {
@@ -115,12 +123,15 @@ const projectDataReducer: LoopReducer<ProjectState, Action> = (
         })
       );
     case getType(projectDataFetchFailure):
-      return {
-        ...state,
-        projectData: {
-          errorMessage: action.payload
-        }
-      };
+      return loop(
+        {
+          ...state,
+          projectData: {
+            errorMessage: action.payload
+          }
+        },
+        Cmd.run(showResourceFailedToast)
+      );
     case getType(staticDataFetchSuccess):
       return {
         ...state,
@@ -129,16 +140,22 @@ const projectDataReducer: LoopReducer<ProjectState, Action> = (
         }
       };
     case getType(staticDataFetchFailure):
-      return {
-        ...state,
-        staticData: {
-          errorMessage: action.payload
-        }
-      };
+      return loop(
+        {
+          ...state,
+          staticData: {
+            errorMessage: action.payload
+          }
+        },
+        Cmd.run(showResourceFailedToast)
+      );
     case getType(updateDistrictsDefinition):
       return "resource" in state.projectData && "resource" in state.staticData
         ? loop(
-            state,
+            {
+              ...state,
+              saving: "saving"
+            },
             Cmd.run(patchProject, {
               successActionCreator: updateDistrictsDefinitionSuccess,
               failActionCreator: updateDistrictsDefinitionFailure,
@@ -157,14 +174,29 @@ const projectDataReducer: LoopReducer<ProjectState, Action> = (
           )
         : state;
     case getType(updateDistrictsDefinitionSuccess):
-      return "resource" in state.projectData
-        ? loop(state, Cmd.action(projectFetch(state.projectData.resource.project.id)))
-        : state;
+      return loop(
+        state,
+        Cmd.run(
+          () => {
+            return fetchProjectGeoJson(action.payload.id).then((geojson: DistrictsGeoJSON) => ({
+              project: action.payload,
+              geojson
+            }));
+          },
+          {
+            successActionCreator: projectFetchSuccess,
+            failActionCreator: updateDistrictsDefinitionFailure
+          }
+        )
+      );
     case getType(updateDistrictsDefinitionFailure):
-      // TODO (#188): implement a status area to display errors for this and other things
-      // eslint-disable-next-line
-      console.log("Error patching districts definition: ", action.payload);
-      return state;
+      return loop(
+        {
+          ...state,
+          saving: "failed"
+        },
+        Cmd.run(showActionFailedToast)
+      );
     default:
       return state as never;
   }